+++ usr.bin/window/lcmd2.c
@@ -156,7 +156,7 @@ strtime(struct timeval *t)
 	char *p = buf;
 
 	if (t->tv_sec > 60*60) {
-		(void) sprintf(p, "%ld:", (long int)t->tv_sec / (60*60));
+		(void) sprintf(p, "%lld:", t->tv_sec / (60*60LL));
 		while (*p++)
 			;
 		p--;
@@ -164,15 +164,15 @@ strtime(struct timeval *t)
 		fill++;
 	}
 	if (t->tv_sec > 60) {
-		(void) sprintf(p, fill ? "%02ld:" : "%ld:", t->tv_sec / 60);
+		(void) sprintf(p, fill ? "%02lld:" : "%lld:", t->tv_sec / 60LL);
 		while (*p++)
 			;
 		p--;
 		t->tv_sec %= 60;
 		fill++;
 	}
-	(void) sprintf(p, fill ? "%02ld.%02ld" : "%ld.%02ld",
-		t->tv_sec, t->tv_usec / 10000);
+	(void) sprintf(p, fill ? "%02lld.%02d" : "%lld.%02d",
+		(long long int)t->tv_sec, (int)(t->tv_usec / 10000));
 	return buf;
 }
