+++ usr.bin/xinstall/xinstall.c
@@ -1098,7 +1098,7 @@ metadata_log(const char *path, const char *type, struct timeval *tv,
 	if (tags)
 		fprintf(metafp, " tags=%s", tags);
 	if (tv != NULL && dopreserve)
-		fprintf(metafp, " time=%ld.%ld", tv[1].tv_sec, tv[1].tv_usec);
+		fprintf(metafp, " time=%lld.%ld", (long long int)tv[1].tv_sec, (long int)tv[1].tv_usec);
 	if (digestresult && digest)
 		fprintf(metafp, " %s=%s", digest, digestresult);
 	fputc('\n', metafp);
